requirements.txt
beautifulsoup4==4.11.2
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
cloudscraper==1.2.68
flake8==6.0.0
idna==3.4
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==2.6.2
pycodestyle==2.10.0
pyflakes==3.0.1
pyparsing==3.0.9
requests==2.28.2
requests-toolbelt==0.10.1
soupsieve==2.3.2.post1
tomli==2.0.1
urllib3==1.26.14
